change isPrimitive to isString
